requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
streamlit==1.44.1
sqlalchemy==2.0.39
pydantic==2.11.2
pydantic-settings==2.8.1
PyYAML==6.0.2
pandas==2.2.3
duckdb==1.2.2
chromadb==0.6.3
numpy==2.2.4
python-dotenv==1.1.0
pytest==8.3.5
httpx==0.28.1
networkx==3.4.2
spacy==3.8.5
keybert==0.8.5
bertopic==0.17.0
transformers==4.50.3
vaderSentiment==3.3.2
faster-whisper==1.1.1
yt-dlp==2025.2.19
instaloader==4.14.1
apify-client==1.7.1
psycopg2-binary==2.9.10
google-api-python-client==2.154.0
python-docx==1.1.2
